2020.09.03 refactoring

+++ model.php
@@ -58,10 +58,10 @@
 		 */
 		public function insert(): bool {
 		    $result = false;
-		    $w = wp_insert_term($this->name, 'product_cat', [
-		        "description" => $this->description,
-		        "parent" => $this->parent
-		    ]);
+   		    $w = wp_insert_term($this->name, 'product_cat', [
+    		        "description" => $this->description,
+    		        "parent" => $this->parent
+   		    ]);
 		    if (isset($w['term_id'])) {
 		        $this->id = $w['term_id'];
 		        $result = true;
@@ -74,35 +74,42 @@
 		
 		/**
 		 * update data in database
+		 * @params bool $onlyMetas
 		 * @return bool
 		 */
-		public function modify(): bool {
-		    $result = false;
-		    $w = wp_update_term($this->id, 'product_cat', [
-		        "name" => $this->name,
-		        "description" => $this->description,
-		        "parent" => $this->parent
-		    ]);
-		    if (isset($w['term_id'])) {
-		        $this->id = $w['term_id'];
-		        $result = true;
-		        foreach ($this->metaFields as $metaField) {
-    		        if (!update_term_meta($this->id, $metaField, $this->$metaField)) {$result = false;}
-		        }
+		public function modify(bool $onlyMetas): bool {
+		    $result = true;
+		    if (!$onlyMetas) {
+    		    $w = wp_update_term($this->id, 'product_cat', [
+    		        "name" => $this->name,
+    		        "description" => $this->description,
+    		        "parent" => $this->parent
+    		    ]);
 		    }
+	        foreach ($this->metaFields as $metaField) {
+   		        if (!update_term_meta($this->id, $metaField, $this->$metaField)) {$result = false;}
+	        }
 		    return $result;
 		}
 		
 		/**
 		 * delete data from database
+		 * @params bool $onlyMetas
 		 * @return bool
 		 */
-		public function remove(): bool {
+		public function remove(bool $onlyMetas): bool {
 		    $result = false;
-		    if (wp_delete_term($this->id, 'product_cat')) {
+		    if (!$onlyMetas) {
+    		    if (wp_delete_term($this->id, 'product_cat')) {
+    		        $result = true;
+    		        foreach ($this->metaFields as $metaField) {
+    		            if (!delete_term_meta($this->id, $metaField)) {$result = false;}
+    		        }
+    		    }
+		    } else {
 		        $result = true;
 		        foreach ($this->metaFields as $metaField) {
-		          if (!delete_term_meta($this->id, $metaField)) {$result = false;}
+		            if (!delete_term_meta($this->id, $metaField)) {$result = false;}
 		        }
 		    }
 		    return $result;
